+++ CHANGELOG.md
@@ -5,6 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.1.0] - 2025-02-?
+
+### Changed
+
+- Can add custom effects from UI
+- Config file is now Json5
+
+### Fixed
+
+- Fix mod crashing because of custom effects
+
 ## [2.0.1] - 2024-01-29
 
 ### Fixed